+++ src/main/java/woowacourse/shoppingcart/dto/LoginResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.dto;
+package woowacourse.shoppingcart.dto;
 
 public class LoginResponse {
