+++ src/main/java/woowacourse/shoppingcart/support/AuthorizationExtractor.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.support;
+package woowacourse.shoppingcart.support;
 
 import java.util.Enumeration;
 import javax.servlet.http.HttpServletRequest;
