+++ src/main/java/woowacourse/shoppingcart/support/Login.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.support;
+package woowacourse.shoppingcart.support;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
