+++ src/main/java/woowacourse/shoppingcart/ui/AuthController.java
@@ -1,13 +1,13 @@
-package woowacourse.auth.ui;
+package woowacourse.shoppingcart.ui;
 
 import javax.validation.Valid;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import woowacourse.auth.application.AuthService;
-import woowacourse.auth.dto.LoginRequest;
-import woowacourse.auth.dto.LoginResponse;
+import woowacourse.shoppingcart.application.AuthService;
+import woowacourse.shoppingcart.dto.LoginRequest;
+import woowacourse.shoppingcart.dto.LoginResponse;
 
 @RestController
 public class AuthController {
